Text: introduce (Default)TextRenderer
